+++ asteroid/filterbanks/free_fb.py
@@ -12,7 +12,7 @@ class FreeFB(Filterbank):
         kernel_size (int): Length of the filters.
         stride (int, optional): Stride of the convolution.
             If None (default), set to ``kernel_size // 2``.
-        sample_rate (int): Sample rate of the expected audio.
+        sample_rate (float): Sample rate of the expected audio.
             Defaults to 8000.
 
     Attributes:
@@ -24,7 +24,7 @@ class FreeFB(Filterbank):
         Antoine Deleforge, Emmanuel Vincent.
     """
 
-    def __init__(self, n_filters, kernel_size, stride=None, sample_rate=8000, **kwargs):
+    def __init__(self, n_filters, kernel_size, stride=None, sample_rate=8000.0, **kwargs):
         super().__init__(n_filters, kernel_size, stride=stride, sample_rate=sample_rate)
         self._filters = nn.Parameter(torch.ones(n_filters, 1, kernel_size))
         for p in self.parameters():
